Move notifications to a standalone file
